feat: implement add game via BGG with tests

+++ apps/api/src/Api/Services/Pdf/PositionedTextExtractionStrategy.cs
@@ -33,9 +33,9 @@ internal sealed class PositionedTextExtractionStrategy : IEventListener
                 var baseline = characterInfo.GetBaseline();
                 var startPoint = baseline.GetStartPoint();
                 var endPoint = baseline.GetEndPoint();
-                var x = (float)startPoint.Get(Vector.I1);
-                var y = (float)startPoint.Get(Vector.I2);
-                var width = (float)Math.Abs(endPoint.Get(Vector.I1) - startPoint.Get(Vector.I1));
+                var x = startPoint.Get(Vector.I1);
+                var y = startPoint.Get(Vector.I2);
+                var width = Math.Abs(endPoint.Get(Vector.I1) - startPoint.Get(Vector.I1));
 
                 if (width <= 0)
                 {
